serverless-ml-intro/requirements.txt
hopsworks
joblib
scikit-learn==0.24.0
seaborn
dataframe-image
modal-client
gradio